requirements.txt
requests==2.22.0
selenium==3.141.0
webdriver_manager==3.2.2
yagmail==0.11.224
beautifulsoup4==4.9.3
python-dotenv==0.14.0
